test: fix and refactored access tests
