+++ tests/app/Filament/ActivityResourceTest.php
@@ -5,6 +5,7 @@ namespace App\Filament;
 use App\BaseFilamentTestCase;
 use App\Filament\AccessCheckingHelpers\ChecksListingFilamentAccess;
 use App\Filament\AccessCheckingHelpers\ChecksViewFilamentAccess;
+;
 use App\Filament\Resources\ActivityResource;
 use App\Models\User\RoleKeys;
 use Illuminate\Database\Eloquent\Model;
