+++ tests/app/Filament/ChecksOperationsContributorAccess.php
@@ -2,8 +2,8 @@
 
 namespace App\Filament;
 
-use App\Filament\AccessCheckingHelpers\ChecksCreateFilamentAccess;
-use App\Filament\AccessCheckingHelpers\ChecksEditFilamentAccess;
+use App\Filament\AccessCheckingHelpers\ChecksCreateOperationsContributorAccess;
+use App\Filament\AccessCheckingHelpers\ChecksEditOperationsContributorAccess;
 use App\Filament\AccessCheckingHelpers\ChecksListingFilamentAccess;
 use App\Filament\AccessCheckingHelpers\ChecksViewFilamentAccess;
 
@@ -11,10 +11,10 @@ use App\Filament\AccessCheckingHelpers\ChecksViewFilamentAccess;
  * Rather than repeating the same tests for every class that follows the default Filament access policy,
  * this trait provides the tests and exposes a few methods to customise them.
  */
-trait ChecksDefaultFilamentAccess
+trait ChecksOperationsContributorAccess
 {
-    use ChecksCreateFilamentAccess;
-    use ChecksEditFilamentAccess;
+    use ChecksCreateOperationsContributorAccess;
+    use ChecksEditOperationsContributorAccess;
     use ChecksListingFilamentAccess;
     use ChecksViewFilamentAccess;
 }
