+++ tests/app/Policies/ActivityLogPolicyTest.php
@@ -33,42 +33,42 @@ class ActivityLogPolicyTest extends BaseFunctionalTestCase
     {
         return [
             'View No Role' => ['view', null, false],
-            'View Operations Contributor' => [RoleKeys::OPERATIONS_CONTRIBUTOR, false],
+            'View Operations Contributor' => ['view',RoleKeys::OPERATIONS_CONTRIBUTOR, false],
             'View Operations' => ['view', RoleKeys::OPERATIONS_TEAM, false],
             'View Web' => ['view', RoleKeys::WEB_TEAM, true],
             'View DSG' => ['view', RoleKeys::DIVISION_STAFF_GROUP, true],
             'View Any No Role' => ['viewAny', null, false],
-            'View Any Operations Contributor' => [RoleKeys::OPERATIONS_CONTRIBUTOR, false],
+            'View Any Operations Contributor' => ['viewAny',RoleKeys::OPERATIONS_CONTRIBUTOR, false],
             'View Any Operations' => ['viewAny', RoleKeys::OPERATIONS_TEAM, false],
             'View Any Web' => ['viewAny', RoleKeys::WEB_TEAM, true],
             'View Any DSG' => ['viewAny', RoleKeys::DIVISION_STAFF_GROUP, true],
             'Update No Role' => ['update', null, false],
-            'Update Any Operations Contributor' => [RoleKeys::OPERATIONS_CONTRIBUTOR, false],
+            'Update Any Operations Contributor' => ['update',RoleKeys::OPERATIONS_CONTRIBUTOR, false],
             'Update Operations' => ['update', RoleKeys::OPERATIONS_TEAM, false],
             'Update Web' => ['update', RoleKeys::WEB_TEAM, false],
             'Update DSG' => ['update', RoleKeys::DIVISION_STAFF_GROUP, false],
             'Create No Role' => ['create', null, false],
-            'Create Any Operations Contributor' => [RoleKeys::OPERATIONS_CONTRIBUTOR, false],
+            'Create Any Operations Contributor' => ['create',RoleKeys::OPERATIONS_CONTRIBUTOR, false],
             'Create Operations' => ['create', RoleKeys::OPERATIONS_TEAM, false],
             'Create Web' => ['create', RoleKeys::WEB_TEAM, false],
             'Create DSG' => ['create', RoleKeys::DIVISION_STAFF_GROUP, false],
             'Delete No Role' => ['delete', null, false],
-            'Delete Operations Contributor' => [RoleKeys::OPERATIONS_CONTRIBUTOR, false],
+            'Delete Operations Contributor' => ['delete',RoleKeys::OPERATIONS_CONTRIBUTOR, false],
             'Delete Operations' => ['delete', RoleKeys::OPERATIONS_TEAM, false],
             'Delete Web' => ['delete', RoleKeys::WEB_TEAM, false],
             'Delete DSG' => ['delete', RoleKeys::DIVISION_STAFF_GROUP, false],
             'Delete Any No Role' => ['delete', null, false],
-            'Delete Any Operations Contributor' => [RoleKeys::OPERATIONS_CONTRIBUTOR, false],
+            'Delete Any Operations Contributor' => ['delete',RoleKeys::OPERATIONS_CONTRIBUTOR, false],
             'Delete Any Operations' => ['delete', RoleKeys::OPERATIONS_TEAM, false],
             'Delete Any Web' => ['delete', RoleKeys::WEB_TEAM, false],
             'Delete Any DSG' => ['delete', RoleKeys::DIVISION_STAFF_GROUP, false],
             'Restore No Role' => ['restore', null, false],
-            'Restore Operations Contributor' => [RoleKeys::OPERATIONS_CONTRIBUTOR, false],
+            'Restore Operations Contributor' => ['restore',RoleKeys::OPERATIONS_CONTRIBUTOR, false],
             'Restore Operations' => ['restore', RoleKeys::OPERATIONS_TEAM, false],
             'Restore Web' => ['restore', RoleKeys::WEB_TEAM, false],
             'Restore DSG' => ['restore', RoleKeys::DIVISION_STAFF_GROUP, false],
             'Force Delete No Role' => ['forceDelete', null, false],
-            'Force Delete Operations Contributor' => [RoleKeys::OPERATIONS_CONTRIBUTOR, false],
+            'Force Delete Operations Contributor' => ['forceDelete',RoleKeys::OPERATIONS_CONTRIBUTOR, false],
             'Force Delete Operations' => ['forceDelete', RoleKeys::OPERATIONS_TEAM, false],
             'Force Delete Web' => ['forceDelete', RoleKeys::WEB_TEAM, false],
             'Force Delete DSG' => ['forceDelete', RoleKeys::DIVISION_STAFF_GROUP, false],
